Prevent errors when cff is passed explicit nils
Before this change, the default on field caches is set to empty string.
This allows calling methods like .empty? on keys that where never explicitly set.
This works well in most cases, however it is really easy to have explictly nil keys which default will not override.
Instead we can handle explicit nils when reading the keys and assure that explicit and implicit nils are treated the same.