Add `Util::transliterate` to normalise towards ASCII.
Note that this doesn't change the encoding of the string, but tries to
convert as many non-ascii characters as it can into ASCII
approximations. This is only really practical for Latin and nearly-Latin
characters, but as we're dealing with academic papers this is probably
OK.

This method is a simplified reimplementation if `I18n::transliterate`.