Always treat dates as Ruby Date objects.
If the date in the YAML is quoted, e.g., `"2023-06-24"`, then the Ruby YAML
parser - quite reasonably - assumed that they *must* be strings, so does not
parse them into Ruby Date objects.

We always want certain fields to be treated as Date objects, so this commit
ensures that those fields always return a Date object, or `''` if the date
is invalid or empty.

Fixes #122.