Turn on underlying support for YAML anchors/aliases.
This commit adds support for using YAML anchors and aliases in CFF files.
With this in place the underlying Ruby YAML library will load and reference
anchors/aliases as expected, and Ruby CFF will process and validate such
files as expected too. Links between anchors/aliases are preserved in line
with expectations.

Fixes #130